requirements.txt
streamlit
google-generativeai
docxtpl
htmldocx
python-docx